@@ -2,7 +2,7 @@ name: clima_data
 description: A starting point for Dart libraries or applications.
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 
 dependencies:
   clima_core:
