@@ -2,7 +2,7 @@ import 'package:clima_ui/widgets/others/link_text_span.dart';
 import 'package:flutter/material.dart';
 
 class ApiKeyInfoDialog extends StatelessWidget {
-  const ApiKeyInfoDialog({Key? key}) : super(key: key);
+  const ApiKeyInfoDialog({super.key});
 
   @override
   Widget build(BuildContext context) {
