@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class ApiKeyResetDialog extends ConsumerWidget {
-  const ApiKeyResetDialog({Key? key}) : super(key: key);
+  const ApiKeyResetDialog({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) => AlertDialog(
