@@ -8,7 +8,7 @@ import 'package:clima_ui/widgets/others/link_text_span.dart';
 import 'package:flutter/material.dart';
 
 class CreditsDialog extends StatelessWidget {
-  const CreditsDialog({Key? key}) : super(key: key);
+  const CreditsDialog({super.key});
 
   @override
   Widget build(BuildContext context) {
