@@ -7,7 +7,7 @@
 import 'package:flutter/material.dart';
 
 class SettingsDivider extends StatelessWidget {
-  const SettingsDivider({Key? key}) : super(key: key);
+  const SettingsDivider({super.key});
 
   @override
   Widget build(BuildContext context) {
