@@ -11,7 +11,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:sizer/sizer.dart';
 
 class MainInfoWidget extends ConsumerWidget {
-  const MainInfoWidget({Key? key}) : super(key: key);
+  const MainInfoWidget({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
