@@ -6,16 +6,13 @@
 
 import 'package:collection/collection.dart';
 
-class DarkThemeModel {
-  const DarkThemeModel._(this.string);
+enum DarkThemeModel {
+  darkGrey('darkGrey'),
+  black('black');
+
+  const DarkThemeModel(this.string);
 
   final String string;
-
-  static const darkGrey = DarkThemeModel._('darkGrey');
-
-  static const black = DarkThemeModel._('black');
-
-  static const values = [darkGrey, black];
 
   static DarkThemeModel parse(String string) {
     final theme = values.firstWhereOrNull((theme) => theme.string == string);
