@@ -6,18 +6,14 @@
 
 import 'package:collection/collection.dart';
 
-class ThemeModel {
-  const ThemeModel._(this.string);
+enum ThemeModel {
+  systemDefault('systemDefault'),
+  light('light'),
+  dark('dark');
+
+  const ThemeModel(this.string);
 
   final String string;
-
-  static const systemDefault = ThemeModel._('systemDefault');
-
-  static const light = ThemeModel._('light');
-
-  static const dark = ThemeModel._('dark');
-
-  static const values = [systemDefault, light, dark];
 
   static ThemeModel parse(String string) {
     final theme = values.firstWhereOrNull((theme) => theme.string == string);
